requirements.txt
PyJWT==2.9.0
Squad1CoreAuthenticator==1.3
Squad1CoreSender==1.2.5
ldap3==2.9.1
pika==1.3.2
pyasn1==0.6.1